Add files via upload
Work in progress powerpoint slides... tried to go with what Dr. Mukherjee said about explaining the importance of your research in this limited amount of time instead of trying to explain everything that you did. There are notes below each slide with ideas on what to say. Make changes as you see fit!