Add files via upload
I couldn't find the one we printed, and so I made the change to "strong a" and fixed the bar graph title. I can't save it as a pdf though because the formula gets all messed up. Could one of you save it as a pdf and either send it to Mitch or upload it to github and I'll send it to Mitch?